Improve static preview command

+++ scripts/staticPreview.js
@@ -3,6 +3,7 @@ const sharp = require('sharp')
 const path = require('path')
 const mkdirp = require('mkdirp')
 const rimraf = require('rimraf')
+const fs = require('fs')
 
 const PAGE_SIZE = 200
 async function fetchDocs(page = 1) {
@@ -17,18 +18,37 @@ async function fetchDocs(page = 1) {
 const VALID_RESIZE_EXTS = ['.jpg', '.jpeg', '.png']
 
 const SIZES = {
-  'small': {
-    height: 100
+  small: {
+    height: 100,
   },
-  'medium': {
-    height: 400
-  }
+  medium: {
+    height: 400,
+  },
 }
 
-async function makeImagePreview(image) {
+async function makeImagePreview(image, clean) {
   const ext = path.extname(image.image)
   // Invalid ext to resize
   if (!VALID_RESIZE_EXTS.includes(ext.toLowerCase())) {
+    return
+  }
+  const fileName = path.basename(image.image)
+
+  let doSizes
+  if (clean) {
+    doSizes = Object.keys(SIZES)
+  } else {
+    doSizes = Object.keys(SIZES).filter((size) => {
+      const outFile = path.join(
+        __dirname,
+        `../public/preview/${size}/${fileName}`
+      )
+      return !fs.existsSync(outFile)
+    })
+  }
+
+  // No need to download image
+  if (doSizes.length === 0) {
     return
   }
 
@@ -37,23 +57,26 @@ async function makeImagePreview(image) {
     .responseType('blob')
     .then((r) => r.body)
 
-  const fileName = path.basename(image.image)
-
-  for (const size of Object.keys(SIZES)) {
+  for (const size of doSizes) {
     await mkdirp(path.join(__dirname, `../public/preview/${size}`))
-    const outFile = path.join(__dirname, `../public/preview/${size}/${fileName}`)
+    const outFile = path.join(
+      __dirname,
+      `../public/preview/${size}/${fileName}`
+    )
     await sharp(imageBlob).resize(SIZES[size]).toFile(outFile)
   }
-  console.log(`${image.image} resized!`)
 }
 
-async function resizeAllImages() {
-  // Remove previuous preview
-  rimraf.sync(path.join(__dirname, `../public/preview`))
+async function resizeAllImages(clean) {
+  // Remove previuous preview (when needed)
+  if (clean) {
+    rimraf.sync(path.join(__dirname, `../public/preview`))
+  }
 
   let page = 1
   let fetchCount = 0
   let count = 0
+  let resized = 0
 
   do {
     const data = await fetchDocs(page)
@@ -63,14 +86,25 @@ async function resizeAllImages() {
 
     for (const doc of results) {
       for (const image of doc.images) {
-        await makeImagePreview(image)
+        await makeImagePreview(image, clean)
+        resized++
+        console.log(`${image.image} resized! ${resized}/${count}`)
       }
     }
     page++
   } while (fetchCount < count)
 }
 
-resizeAllImages().then(
+const { program } = require('commander');
+
+program
+  .option('-c, --clean', 'Clean previous preview')
+
+program.parse(process.argv);
+
+const options = program.opts();
+
+resizeAllImages(options.clean ?? false).then(
   () => {
     console.log('Done!')
   },
